Move JS file to src for wp-scripts support
